+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [5.0.3]
+
+### Fixed
+- Added field validation to the output of the error msg.
+- Additional fixes for legacy output data.
+
 ## [5.0.2]
 
 ### Added
@@ -597,6 +603,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[5.0.2]: https://github.com/infinum/eightshift-forms/compare/5.0.2...5.0.3
 [5.0.1]: https://github.com/infinum/eightshift-forms/compare/5.0.1...5.0.2
 [5.0.1]: https://github.com/infinum/eightshift-forms/compare/5.0.0...5.0.1
 [5.0.0]: https://github.com/infinum/eightshift-forms/compare/4.0.7...5.0.0